+++ preprocess_web/code/ravens_metadata_apps/preprocess_jobs/templates/preprocess/homepage.html
@@ -98,7 +98,7 @@
           <div class="col-lg-4">
           <img class="rounded-circle" src="{% static "icon/edit.png" %}" alt="Generic placeholder image" width="60" height="60"style="background-color: #FAFAFA">
           <h2>Lists</h2>
-          <p>Get the list of all the preprocess jobs.  <p><a class="btn btn-primary" href="{% url 'view_list' %}" style="background-color: #ffcccc" role="button">Check List</a></p>
+          <p>Get the list of all the preprocess jobs.  <p><a class="btn btn-primary" href="{% url 'view_job_list' %}" style="background-color: #ffcccc" role="button">Check List</a></p>
 
           </div><!-- /.col-lg-4 -->
