+++ preprocess_web/code/ravens_metadata_apps/preprocess_jobs/templates/preprocess/preprocess-job-detail.html
@@ -2,7 +2,7 @@
 {% block main_content %}
 {% if iterable %}
 <h4 class="border-bottom border-gray pb-2 mb-0" style="text-align: center">{{ name.name }}</h4>
-    <a class="btn btn-info" href="{%  url 'view_list' %}" style="float: left">Back</a>
+    <a class="btn btn-info" href="{%  url 'view_job_list' %}" style="float: left">Back</a>
 <br>
 <br>
 <table class="table table-hover">
